requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
gradio>=3.0.0
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
pillow>=8.0.0
requests>=2.25.0
opencv-python>=4.5.0
matplotlib>=3.4.0
